refactor: seperate product and cart services
